+++ plugins/builtin/include/content/providers/memory_file_provider.hpp
@@ -23,9 +23,9 @@ namespace hex::plugin::builtin {
         void writeRaw(u64 offset, const void *buffer, size_t size) override;
         [[nodiscard]] u64 getActualSize() const override { return this->m_data.size(); }
 
-        void resizeRaw(size_t newSize) override;
-        void insertRaw(u64 offset, size_t size) override;
-        void removeRaw(u64 offset, size_t size) override;
+        void resizeRaw(u64 newSize) override;
+        void insertRaw(u64 offset, u64 size) override;
+        void removeRaw(u64 offset, u64 size) override;
 
         void save() override;
